backend/requirements.txt
annotated-types==0.6.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-PageDown==0.4.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
